requirements.txt
pystray
Pillow
hidapi
blynclight
pywebview
pyinstaller
